renaming old implementation to _wbx before adding the ported one
