+++ src/BizHawk.Emulation.Cores/Consoles/Sony/PSP_WBX/PPSSPP_WBX.IDriveLight.cs
@@ -1,8 +1,8 @@
 using BizHawk.Emulation.Common;
 
-namespace BizHawk.Emulation.Consoles.Sony.PSP
+namespace BizHawk.Emulation.Consoles.Sony.PSP_WBX
 {
-	public partial class PPSSPP : IDriveLight
+	public partial class PPSSPP_WBX : IDriveLight
 	{
 		public bool DriveLightEnabled { get; private set; }
 		public bool DriveLightOn { get; private set; }
